+++ src/main/java/com/global/api/mapping/GpApiMapping.java
@@ -52,6 +52,7 @@ public class GpApiMapping {
     private static final String DOCUMENT_UPLOAD = "DOCUMENT_UPLOAD";
     private static final String FILE_CREATE = "FILE_CREATE";
     private static final String FILE_SINGLE = "FILE_SINGLE";
+    public static final String BLIK = "blik";
 
     public static Transaction mapResponse(String rawResponse) throws GatewayException {
         Transaction transaction = new Transaction();
@@ -135,7 +136,6 @@ public class GpApiMapping {
             BatchSummary batchSummary = new BatchSummary();
             batchSummary.setBatchReference(json.getString("batch_id"));
             transaction.setBatchSummary(batchSummary);
-
             if (json.has("payment_method")) {
                 JsonDoc paymentMethod = json.get("payment_method");
 
@@ -207,6 +207,9 @@ public class GpApiMapping {
                     transaction.setPaymentMethodType(PaymentMethodType.ACH);
                 } else if (paymentMethod.has("apm")) {
                     transaction.setPaymentMethodType(PaymentMethodType.APM);
+                    if (paymentMethod.get("apm").getString("provider") != null && paymentMethod.get("apm").getString("provider").equalsIgnoreCase(BLIK)) {
+                        mapAPMResponse(json, transaction);
+                    }
                 }
             }
 
@@ -596,6 +599,12 @@ public class GpApiMapping {
         bnplResponse.setRedirectUrl(response.get("payment_method").getString("redirect_url"));
         bnplResponse.setProviderName(response.get("payment_method").get("bnpl").getString("provider"));
         transaction.setBNPLResponse(bnplResponse);
+    }
+    private static void mapAPMResponse(JsonDoc response, Transaction transaction) {
+        var alternativePaymentResponse = new AlternativePaymentResponse();
+        alternativePaymentResponse.setRedirectUrl(response.get("payment_method").getString("redirect_url"));
+        alternativePaymentResponse.setProviderName(response.get("payment_method").get("apm").getString("provider"));
+        transaction.setAlternativePaymentResponse(alternativePaymentResponse);
     }
 
     private static TransactionSummary createTransactionSummary(JsonDoc doc) throws GatewayException {